initial changes

+++ inversion/training/volumetric_rendering/renderer.py
@@ -26,8 +26,8 @@ def project_onto_planes(planes, coordinates):
     # returns projections of shape N*n_planes, M, 2
     N, M, C = coordinates.shape
     n_planes, _, _ = planes.shape
-    coordinates = coordinates.unsqueeze(1).expand(-1, n_planes, -1, -1).reshape(N*n_planes, M, 3)
-    inv_planes = torch.linalg.inv(planes).unsqueeze(0).expand(N, -1, -1, -1).reshape(N*n_planes, 3, 3)
+    coordinates = coordinates.unsqueeze(1).expand(-1, n_planes, -1, -1).reshape(N*n_planes, M, 3).float()
+    inv_planes = torch.linalg.inv(planes).unsqueeze(0).expand(N, -1, -1, -1).reshape(N*n_planes, 3, 3).float()
     projections = torch.bmm(coordinates, inv_planes)
     return projections[..., :2]
